resolved comments removed new file
