+++ src/components/Search/CompareOverTime.tsx
@@ -2,11 +2,11 @@ import { useState } from 'react';
 
 import { Grid, Typography } from '@mui/material';
 import Divider from '@mui/material/Divider';
+import { VariantType, useSnackbar } from 'notistack';
 import { Form } from 'react-router-dom';
 import { style } from 'typestyle';
 
 import { useAppSelector } from '../../hooks/app';
-import useSearchResults from '../../hooks/useSearchResults';
 import { Strings } from '../../resources/Strings';
 import { CompareCardsStyles, SearchStyles, Spacing } from '../../styles';
 import { Changeset, Repository } from '../../types/state';
@@ -29,6 +29,7 @@ function CompareOverTime({
   hasNonEditableState,
   newRevs,
 }: CompareWithTimeProps) {
+  const { enqueueSnackbar } = useSnackbar();
   const [expanded, setExpanded] = useState(false);
   const [timeRangeValue, setTimeRangeValue] = useState(
     86400 as TimeRange['value'],
@@ -38,11 +39,6 @@ function CompareOverTime({
 
   const [repository, setRepository] = useState('try' as Repository['name']);
 
-  const {
-    handleRemoveNewRevision,
-    handleItemToggleInChangesetList,
-    isMaxRevisions,
-  } = useSearchResults();
   const mode = useAppSelector((state) => state.theme.mode);
   const styles = CompareCardsStyles(mode);
   const dropDownStyles = SearchStyles(mode);
@@ -82,22 +78,60 @@ function CompareOverTime({
     setExpanded(!expanded);
   };
 
+  const handleRemoveRevision = (item: Changeset) => {
+    // Currently item seems to be the same object than the one stored in
+    // newInProgressRevs, but it might change in the future. That's why we're
+    // comparing the ids instead of using indexOf directly.
+    const indexInNewChangesets = inProgressRevs.findIndex(
+      (rev) => rev.id === item.id,
+    );
+    const revisionsNew = [...inProgressRevs];
+    revisionsNew.splice(indexInNewChangesets, 1);
+    setInProgressRevs(revisionsNew);
+  };
+
+  const handleItemToggleInChangesetList = ({
+    item,
+    changesets,
+  }: {
+    item: Changeset;
+    changesets: Changeset[];
+  }) => {
+    // Warning: `item` isn't always the same object than the one in
+    // `changesets`, therefore we need to compare the id. This happens when the
+    // data in `changesets` comes from the loader, but `item` comes from the
+    // search results.
+    const indexInCheckedChangesets = changesets.findIndex(
+      (rev) => rev.id === item.id,
+    );
+    const isChecked = indexInCheckedChangesets >= 0;
+    const newChecked = [...changesets];
+
+    // if item is not already checked, add to checked
+    if (!isChecked) {
+      newChecked.push(item);
+    } else if (isChecked) {
+      // if item is already checked, remove from checked
+      newChecked.splice(indexInCheckedChangesets, 1);
+    }
+
+    return newChecked;
+  };
+
   const handleSearchResultsToggle = (item: Changeset) => {
-    const updatedRevs = handleItemToggleInChangesetList({
+    const newNewRevs = handleItemToggleInChangesetList({
       item,
       changesets: inProgressRevs,
     });
 
-    if (isMaxRevisions(updatedRevs)) {
+    const maxRevisions = 3;
+    if (newNewRevs.length > maxRevisions) {
+      const variant: VariantType = 'warning';
+      enqueueSnackbar(`Maximum ${maxRevisions} revisions.`, { variant });
       return;
     }
-
-    setInProgressRevs(updatedRevs);
-  };
-
-  const handleRemoveRevision = (item: Changeset) => {
-    const updatedRevs = handleRemoveNewRevision(item, inProgressRevs);
-    setInProgressRevs(updatedRevs);
+    // if there are already `maxRevisions` checked revisions, print a warning
+    setInProgressRevs(newNewRevs);
   };
 
   return (
@@ -149,6 +183,8 @@ function CompareOverTime({
             </div>
 
             <CompareButton
+              //this disabled btn is temporary until
+              //fetching the data is implemented
               isDisabled={true}
               label={strings.sharedCollasped.button}
             />
