+++ src/components/Search/CompareWithBase.tsx
@@ -3,12 +3,11 @@ import { useState } from 'react';
 import { Typography } from '@mui/material';
 import Divider from '@mui/material/Divider';
 import Grid from '@mui/material/Grid';
-import { useSnackbar } from 'notistack';
+import { VariantType, useSnackbar } from 'notistack';
 import { Form } from 'react-router-dom';
 import { style } from 'typestyle';
 
 import { useAppSelector } from '../../hooks/app';
-import useSearchResults from '../../hooks/useSearchResults';
 import { Strings } from '../../resources/Strings';
 import { CompareCardsStyles, SearchStyles, Spacing } from '../../styles';
 import type { Changeset, Repository } from '../../types/state';
@@ -69,11 +68,6 @@ function CompareWithBase({
 }: CompareWithBaseProps) {
   const [expanded, setExpanded] = useState(true);
   const { enqueueSnackbar } = useSnackbar();
-  const {
-    handleRemoveNewRevision,
-    handleItemToggleInChangesetList,
-    isMaxRevisions,
-  } = useSearchResults();
 
   // The "committed" base and new revisions initialize the staging state.
   // These states are snapshots of selections, that the user can either save (by
@@ -157,6 +151,34 @@ function CompareWithBase({
     setInProgressNewRevs(newStagingRevs);
   };
 
+  const handleItemToggleInChangesetList = ({
+    item,
+    changesets,
+  }: {
+    item: Changeset;
+    changesets: Changeset[];
+  }) => {
+    // Warning: `item` isn't always the same object than the one in
+    // `changesets`, therefore we need to compare the id. This happens when the
+    // data in `changesets` comes from the loader, but `item` comes from the
+    // search results.
+    const indexInCheckedChangesets = changesets.findIndex(
+      (rev) => rev.id === item.id,
+    );
+    const isChecked = indexInCheckedChangesets >= 0;
+    const newChecked = [...changesets];
+
+    // if item is not already checked, add to checked
+    if (!isChecked) {
+      newChecked.push(item);
+    } else if (isChecked) {
+      // if item is already checked, remove from checked
+      newChecked.splice(indexInCheckedChangesets, 1);
+    }
+
+    return newChecked;
+  };
+
   const handleRemoveRevisionBase = (item: Changeset) => {
     // Currently item seems to be the same object than the one stored in
     // baseInProgressRevs, but it might change in the future. That's why we're
@@ -177,8 +199,15 @@ function CompareWithBase({
   };
 
   const handleRemoveRevisionNew = (item: Changeset) => {
-    const updatedRevs = handleRemoveNewRevision(item, newInProgressRevs);
-    setInProgressNewRevs(updatedRevs);
+    // Currently item seems to be the same object than the one stored in
+    // newInProgressRevs, but it might change in the future. That's why we're
+    // comparing the ids instead of using indexOf directly.
+    const indexInNewChangesets = newInProgressRevs.findIndex(
+      (rev) => rev.id === item.id,
+    );
+    const revisionsNew = [...newInProgressRevs];
+    revisionsNew.splice(indexInNewChangesets, 1);
+    setInProgressNewRevs(revisionsNew);
   };
 
   const handleSearchResultsToggleBase = (item: Changeset) => {
@@ -195,11 +224,13 @@ function CompareWithBase({
       changesets: newInProgressRevs,
     });
 
-    if (isMaxRevisions(newNewRevs)) {
+    const maxRevisions = 3;
+    if (newNewRevs.length > maxRevisions) {
+      const variant: VariantType = 'warning';
+      enqueueSnackbar(`Maximum ${maxRevisions} revisions.`, { variant });
       return;
     }
     // if there are already `maxRevisions` checked revisions, print a warning
-
     setInProgressNewRevs(newNewRevs);
   };
